Add files via upload
This is my code for the Udacity FrontEnd web developer course project on Pixel Art Maker.
The codes written in design.js are solely my intellectual properties and as such needs my permissions before it can be edited.

This Project was to produce a grid using the number of columns and rows inputted by the user and so to design specific boxes in the grid by given them different background colors according to user input. 

Further description of the code functions are given within the design.js file, feel free to learn from it.